review fixes made
