+++ DonjinKrawler-Server/src/main/java/donjinkrawler/enemies/Chicken.java
@@ -1,4 +1,4 @@
-package donjinkrawler.enemy_generation;
+package donjinkrawler.enemies;
 
 public class Chicken extends Enemy {
