+++ DonjinKrawler-Server/src/main/java/donjinkrawler/enemies/EnemyType.java
@@ -1,0 +1,7 @@
+package donjinkrawler.enemies;
+
+public enum EnemyType {
+    CHICKEN,
+    SKELETON,
+    ZOMBIE;
+}
